+++ src/js/ui/NewCharting/ToggleSwitch.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Switch from "@material-ui/core/Switch";
 import Grid from "@material-ui/core/Grid";
 
-export default function toggleSwitch(props){
+export default function ToggleSwitch(props){
     return (
         <div><Grid component="label" container alignItems="center" spacing={1}>
             <Grid item>Tract</Grid>
